Improve resolver code organization
