+++ parcel-resolver-thematics/stringify-yml-func.js
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
 const path = require('path');
 
 /**
